move sorting picker to a separate module
