+++ lib/core/sorting-picker/sorting-picker.component.spec.ts
