+++ lib/core/sorting-picker/sorting-picker.component.ts
